Add files via upload
Created and devised an algorithm for reading in a 50 million word text file. Then, hashing each string from the text file into a set number of arrays, filled with Linked Lists and Node pointers. This is all done in constant (O of 1) time. Searching, adding, and rehashing is also constant time.